+++ ideascube/widgets.py
@@ -1,0 +1,10 @@
+from django.forms import widgets
+
+
+class LangSelect(widgets.Select):
+
+    def render_option(self, selected_choices, option_value, option_label):
+        if option_value:
+            option_label = '{} ({})'.format(option_label, option_value)
+        return super().render_option(selected_choices, option_value,
+                                     option_label)
